Add feature for including openssl-sys
Added a new feature (openssl-static) that includes the openssl-sys crate
to fix static builds.

Renamed src/lib.rs to src/bindgen.rs and added a new lib.rs that
re-exports everything from bindgen.rs to prevent future bindgen runs
from reverting this change.

This fixes #25
